Move remaining files to diagnostics namespace
